WARN-1: match the scanf destination type to its conversion
Eight -Wformat warnings: sscanf(..., "%x", &temp) where temp is an unsigned long.
%x writes an unsigned INT -- four bytes -- so into an eight-byte destination it
leaves the upper half untouched.

It is correct today only by coincidence: temp is zero-initialised, the values are
32-bit radio-subtitle colours, and x86-64 is little-endian so the four bytes
written are the significant ones. Retyped to unsigned int to match the
conversion; the assignments out of it widen as before.

+++ src/sim/otwdrive/radiosubtitle.cpp
@@ -635,7 +635,11 @@ void RadioSubTitle::SetChannelColours(char* flight, char* toPackage, char* ToFro
     unsigned long TowerCol = 0;
     unsigned long StandardCol = 0;
 
-    unsigned long temp = 0;
+    // FF_LINUX: %x writes an unsigned INT (4 bytes). Into an 8-byte unsigned
+    // long that leaves the upper half untouched -- it happens to be correct here
+    // only because temp is zero-initialised, the values are 32-bit colours, and
+    // x86-64 is little-endian. Match the type to the conversion.
+    unsigned int temp = 0;
 
     if (flight)
         if (sscanf(flight, "%x", &temp) == 1)
